trunk:
configure: added example/test_one.sh and test_more.sh to configure output.
example/makefile.*: -ldl library in ipopt test case, output test_one.sh.in.
*/test_one.sh: moved to */test_one.sh.in.
*/test_one.sh.in: moved from */test_one.sh.
base_adolc.hpp: correct compliation errors with g++ 4.3.
gpl_version.sh: make version a command line argument (for release versions).
test_more/makefile.*: distribute test_one.sh.in (instead of test_one.sh).
*/makefile.in: new version of automake.
configure.ac: added example/test_one.sh and test_more.sh to configure output.
dos_format.sh: make version a command line argument (so like gpl_license.sh).
build.sh: change add-on prefixes, make test_one.sh executable, better make log.
install_unix_omh.in: new location for Fadbad home page, fix for adolc.
fadbad/*.cpp: new version of Fadbad puts its symbols in fadbad namespace.
adolc/*.cpp: avoid warning in new version of g++. 



git-svn-id: https://projects.coin-or.org/svn/CppAD/trunk@1243 0e562018-b8fb-0310-a933-ecf073c7c197

+++ dos_format.sh
@@ -9,16 +9,17 @@
 # A copy of this license is included in the COPYING file of this distribution.
 # Please visit http://www.coin-or.org/CppAD/ for information on other licenses.
 # -----------------------------------------------------------------------------
+version="$1"
+if [ "$version" == "" ]
+then
+	echo "usage: ./gpl_license.sh version"
+	exit 1
+fi
 #
 if [ -e DosFormat.log ]
 then
 	rm DosFormat.log
 fi
-#
-# get version currently in configure.ac file
-# (in a way that works when version is not a date)
-version=`grep "^ *AC_INIT(" configure.ac | \
-	sed -e 's/[^,]*, *\([^ ,]*\).*/\1/'`
 #
 # delete old version of *.cpl.zip and *.gpl.zip
 for file in cppad-$version.cpl.zip cppad-$version.gpl.zip
